use password field + copy to clipboard button
fixes: #2512
