+++ src/clients/credentials/ClientSecret.tsx
@@ -2,13 +2,16 @@ import React from "react";
 import { useTranslation } from "react-i18next";
 import {
   Button,
-  ClipboardCopy,
   FormGroup,
+  InputGroup,
   Split,
   SplitItem,
 } from "@patternfly/react-core";
+
 import type ClientRepresentation from "@keycloak/keycloak-admin-client/lib/defs/clientRepresentation";
 import type { UseFormMethods } from "react-hook-form";
+import { PasswordInput } from "../../components/password-input/PasswordInput";
+import { CopyToClipboardButton } from "../scopes/CopyToClipboardButton";
 
 export type ClientSecretProps = {
   secret: string;
@@ -23,9 +26,14 @@ export const ClientSecret = ({ secret, toggle, form }: ClientSecretProps) => {
     <FormGroup label={t("clientSecret")} fieldId="kc-client-secret">
       <Split hasGutter>
         <SplitItem isFilled>
-          <ClipboardCopy id="kc-client-secret" isReadOnly>
-            {secret}
-          </ClipboardCopy>
+          <InputGroup>
+            <PasswordInput id="kc-client-secret" value={secret} isReadOnly />
+            <CopyToClipboardButton
+              text={secret}
+              label="clientSecret"
+              variant="control"
+            />
+          </InputGroup>
         </SplitItem>
         <SplitItem>
           <Button
